renamed Product to ProductSpesific and moved all route exports to a barrel file
