+++ src/routes/CheckoutSuccess/index.jsx
@@ -2,7 +2,7 @@ import { Helmet, HelmetProvider } from "react-helmet-async";
 import styles from "./CheckoutSuccess.module.css";
 import Summary from "../../components/Products/ProductsSummary";
 
-export default function CheckoutSuccess() {
+export function CheckoutSuccess() {
   const purchasedProducts = JSON.parse(sessionStorage.getItem("receipt"));
   let paidTotal;
