+++ src/routes/index.jsx
@@ -1,0 +1,7 @@
+export { Checkout } from "./Checkout";
+export { CheckoutSuccess } from "./CheckoutSuccess";
+export { Contact } from "./Contact";
+export { Home } from "./Home";
+export { ProductsList } from "./ProductsList";
+export { ProductSpesific } from "./ProductSpesific";
+export { RouteNotFound } from "./RouteNotFound";
